For a persistent reduction, we generate 2 flavor of 'equivalant' kernels at the same time
- persistent reduction
- regular reduction

A MultiKernel wraps these 2 kernels and pick the one with better performance at runtime.

Here I talk more about implementation details:
- Inductor maintains states for generating kernels. E.g. the wrapper code.  After we generate code for one kernel, we need restore the inductor state before we can generate the counterpart.

***There is one thing I need some comments from others***:
There is one tricky thing about kernel arguments. In general, inductor removes a buffer from the argument list if it's only used inside the kernel.  But somehow a buffer removed by persistent reduction kernel may still be kept by the regular (non-persistent) reduction kernel because of some CSE invalidation rule. My current implementation avoid removing buffers if multi_kernel is enabled. This makes sure both flavors of reduction has consistent argument list.  Another idea I have is, we generate the multi-kernel definition with the union of arguments from both sub-kernels. Let each sub-kernel pick the subset of arguments it wants. But this will make the code-gen or multi-kernel much complex.

I'm not sure if there is some easy and clean way to resolve this.


Testing command:
```

TORCHINDUCTOR_MULTI_KERNEL=1 TORCH_LOGS=+torch._inductor.graph TORCHINDUCTOR_UNIQUE_KERNEL_NAMES=1 python benchmarks/dynamo/huggingface.py --backend inductor --amp --performance --only BertForMaskedLM --training

```


cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx peterbell10 ipiszy yf225 chenyang78 kadeng muchulee8 aakhundov ColinPeppler Xia-Weiwen ngimel

[ghstack-poisoned]

+++ aten/src/ATen/WrapDimUtilsMulti.h
@@ -15,9 +15,9 @@ constexpr size_t dim_bitset_size = 64;
 
 static inline std::bitset<dim_bitset_size> dim_list_to_bitset(
     OptionalIntArrayRef opt_dims,
-    int64_t ndims) {
+    size_t ndims) {
   TORCH_CHECK(
-      ndims <= (int64_t)dim_bitset_size,
+      ndims <= dim_bitset_size,
       "only tensors with up to ",
       dim_bitset_size,
       " dims are supported");
@@ -25,7 +25,7 @@ static inline std::bitset<dim_bitset_size> dim_list_to_bitset(
   if (opt_dims.has_value()) {
     auto dims = opt_dims.value();
     for (const auto i : c10::irange(dims.size())) {
-      size_t dim = maybe_wrap_dim(dims[i], ndims);
+      size_t dim = maybe_wrap_dim(dims[i], static_cast<int64_t>(ndims));
       TORCH_CHECK(
           !seen[dim],
           "dim ",
@@ -34,7 +34,7 @@ static inline std::bitset<dim_bitset_size> dim_list_to_bitset(
       seen[dim] = true;
     }
   } else {
-    for (int64_t dim = 0; dim < ndims; dim++) {
+    for (size_t dim = 0; dim < ndims; dim++) {
       seen[dim] = true;
     }
   }
