+++ torch/csrc/autograd/cpp_hook.h
@@ -10,21 +10,21 @@ using hooks_list =
     std::vector<std::function<at::TensorBase(const at::TensorBase&)>>;
 
 struct CppFunctionTensorPreHook : public FunctionPreHook {
-  CppFunctionTensorPreHook(std::shared_ptr<hooks_list> hooks, int value_idx);
+  CppFunctionTensorPreHook(std::shared_ptr<hooks_list> hooks, size_t value_idx);
   variable_list operator()(const variable_list& values) override;
 
   std::shared_ptr<hooks_list> hooks_;
-  int value_idx_;
+  size_t value_idx_;
 };
 
 struct CppFunctionSingleTensorPreHook : public FunctionPreHook {
   CppFunctionSingleTensorPreHook(
       std::function<at::TensorBase(const at::TensorBase&)> hook,
-      int value_idx);
+      size_t value_idx);
   variable_list operator()(const variable_list& values) override;
 
   std::function<at::TensorBase(const at::TensorBase&)> hook_;
-  int value_idx_;
+  size_t value_idx_;
 };
 
 } // namespace autograd
